+++ src/commands/command/outputs/index.ts
@@ -4,3 +4,4 @@ export type { DeleteTableOutput } from "./delete-table.js";
 export type { GetItemOutput } from "./get-item.js";
 export type { PutItemOutput } from "./put-item.js";
 export type { QueryOutput } from "./query.js";
+export type { UpdateItemOutput } from "./update-item.js";