+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.59.0] - 2026-02-20
+
+### Added
+
+- The `UpdateItem` command and related output `UpdateItemOutput`.
+
+### Changed
+
+- Renamed `UpdateItemParams` to `UpdateItemInput`.
+- Updating items now return `UpdateItemOutput`, an unusable empty object, instead of `void`.
+
 ## [0.58.0] - 2026-02-20
 
 ### Added
@@ -746,6 +757,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.59.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.58.0...v0.59.0
 [0.58.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.57.0...v0.58.0
 [0.57.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.56.0...v0.57.0
 [0.56.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.55.0...v0.56.0